Initial build script for RHEL 6.
We're addressing one last stumbling block: RHEL stores its include and lib files in an unusual directory naming scheme, and we need to parse around this.  Owing to this, the cythonizing stage of dedalus2 itself currently fails.  Should be fixed shortly.

Thanks to Jeff Oishi for a lot of debugging on getting RHEL up and running.  Thanks also to Jim Mckown at JILA for his help.